refactor extracting plug-in code added directly to rembg script file and moving rembg to a folder to treat it like a package
This is a refactor to separate plug-in code that was added into rembg files that makes it difficult to upgrade rembg.
At the creation time of this plug-in bg.py (a file in the 3rd party rembg package) was renamed to rembg.py, and had plug-in code added directly to it. This makes it difficult to grab new versions of rembg, as bg.py changes would have to be merged into this new rembg.py file every time. Having the script in the same file is unnecessary, so it is now separated.
Also, all of the rembg files were moved into a rembg subfolder so it is back to being it's own package. Now new rembg versions can just be dropped into that folder and we just need to edit our own rembg.py file to work with the new version.