Upgraded rembg package to latest, and api use was fixed
The rembg package has been moved to the latest version. Support for new features was added, including 12 models and mask postprocessing. The alpha_matting parameter was also switched to be a bool instead of an int to match the rembg api, and the alpha matting option in the plugin dialog was switched to a checkbox to match.